pyproject.toml
[project]
name = "read-google-sheet"
version = "0.1.0"
description = "Reading public Google Sheets into polars_result"
readme = "README.md"
requires-python = ">=3.14"
license = { text = "MIT" }
authors = [{ name = "Attica-oss", email = "g.mounac@gmail.com" }]
keywords = ["polars", "Google Sheets", "LazyFrame"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.14",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]
dependencies = [
    "polars>=1.0",
    "polars-result>=0.2.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]

[project.urls]
Homepage = "https://github.com/Attica-OSS/read_google_sheet"
Repository = "https://github.com/Attica-OSS/read_google_sheet"
Issues = "https://github.com/Attica-OSS/read_google_sheet/issues"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.11",
    "ty>=0.0.1a7",
    "auto-mix-prep>=0.2.0",
    "marimo>=0.19.11",
    "types-requests>=2.32.4.20260107",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# ---------- ruff ----------
[tool.ruff]
target-version = "py312"
line-length = 99
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # pyflakes
    "I",     # isort
    "N",     # pep8-naming
    "UP",    # pyupgrade
    "ANN",   # flake8-annotations
    "B",     # flake8-bugbear
    "SIM",   # flake8-simplify
    "RUF",   # ruff-specific rules
]
ignore = [
    "ANN401",  # dynamically typed expressions (typing.Any)
]

[tool.ruff.lint.isort]
known-first-party = ["read_google_sheet"]

# ---------- ty ----------
[tool.ty.environment]
python-version = "3.14"

# ---------- pytest ----------
[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "-ra",
]

# ---------- coverage ----------
[tool.coverage.run]
source = ["read_google_sheet"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
